Revert "feat: share post to an Author"
