+++ server/inbox/views.py
@@ -50,13 +50,16 @@ class InboxView(APIView):
             request_author_id = self.kwargs['author_id']
             # TODO: allow sharing of friend's post
             try:
-                Inbox.send_to_inbox(self, self.request.user.id, post_id)
+                a_post = get_object_or_404(Post, pk=post_id, unlisted=False)
             except ValidationError:
                 return Response(f'{post_id} is not a valid UUID.',
                                 status=status.HTTP_400_BAD_REQUEST)
-            except (Post.DoesNotExist, Inbox.DoesNotExist) as e:
-                return Response('Post or Author not found!',
-                                status=status.HTTP_404_NOT_FOUND)
+            data = PostSerializer(a_post).data
+            inbox = get_object_or_404(Inbox, author=Author.objects
+                                      .get(id=self.request.user.id))
+            data['categories'] = list(data['categories'])
+            inbox.items.append(data)
+            inbox.save()
             return Response(f'Shared {post_id} with {request_author_id}',
                             status=status.HTTP_200_OK)
         else:
