+++ server/posts/urls.py
@@ -7,6 +7,5 @@ app_name = 'posts'
 urlpatterns = [
     path('author/<uuid:author_id>/posts/', views.CreatePostView.as_view(), name='create'),
     path('author/<uuid:author_id>/posts/<uuid:pk>/', views.UpdatePostView.as_view(), name='update'),
-    path('author/<uuid:author_id>/posts/<uuid:pk>/share/', views.SharePostView.as_view(), name='share'),
     path('public/', views.PublicPostView.as_view(), name='public'),
 ]
